ran code cleaner

+++ src/commandParsing/mathCommandParsing/ArcTangent.java
@@ -2,22 +2,22 @@ package commandParsing.mathCommandParsing;
 
 import java.util.List;
 import java.util.Queue;
-
 import workspaceState.WorkspaceState;
 import commandParsing.exceptions.RunTimeDivideByZeroException;
 import commandParsing.floatCommandParsing.OneInputFloatCommandParser;
 import drawableobject.DrawableObject;
 
+
 public class ArcTangent extends OneInputFloatCommandParser {
 
-	public ArcTangent(WorkspaceState someWorkspace) {
-		super(someWorkspace);
-	}
+    public ArcTangent (WorkspaceState someWorkspace) {
+        super(someWorkspace);
+    }
 
-	@Override
-	protected double operateOnComponents(List<Double> components, Queue<DrawableObject> objectQueue)
-			throws RunTimeDivideByZeroException {
-		return (double) Math.atan(components.get(0) / (180 / Math.PI));
-	}
+    @Override
+    protected double operateOnComponents (List<Double> components, Queue<DrawableObject> objectQueue)
+            throws RunTimeDivideByZeroException {
+        return Math.atan(components.get(0) / (180 / Math.PI));
+    }
 
 }
